requirements.txt
amqp==5.1.1
APScheduler==3.10.4
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
boto3==1.28.35
botocore==1.31.35
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2023.1
Django==4.2.4
django-auto-logout==0.5.1
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-db-logger==0.1.12
django-editorjs==0.2.1
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-storages==1.13.2
django4-background-tasks==1.2.9
gunicorn==21.2.0
jmespath==1.0.1
kombu==5.3.1
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.6.2
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
